1. ELB is hashable via simplified form 2. elbclean Identity check is very fast
